+++ utils/Logger.h
@@ -4,6 +4,6 @@
 void logState(int time, int processId, enum LogState state, int arrivalTime,
             int runTime, int remainingTime, int waitingTime);
 void logFinished(int time, int processId, enum LogState state, int arrivalTime,
-            int runTime, int remainingTime, int waitingTime, float turnaroundTime, float weightedTurnaroundTime);
+            int runTime, int remainingTime, int waitingTime, int turnaroundTime, float weightedTurnaroundTime);
 
 void logPerformance(float cpuUtilization, float avgWeightedTurnaroundTime, float avgTurnaroundTime);
